Final Code Snippets
Additions to the gui.m file are to output a time-date stamped matrix output for both the sinogram and reconstructed images.

The WaterCal.mat file is a phantom of a cylinder of water within air.

The PhantomCTP404.m file is a code file to recreate the CTP404 module of the Catphan 600 Phantom. This file needs to be run to generate the phantom as the phantom file is too large to upload! This phantom is the basis of the verification for material identification accuracy.

SourceCTP404_80kvp.mat and SourceCTP404_135kvp.mat are source beam spectra which are designed to replicate the beam of a Aquilion ONE CT Scanner at 80kvp and 135kvp respectively.